FastAPIDemo/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pillow
websockets
